URL combining script
After examining the story-level reports that were generated directly from the queries, we noticed a number of articles had multiple titles. These stories were all linked to the same unique URL, but each title had its own entry in the report. To improve count accuracy and clean these results, we run an SQL script over the BigQuery report to combine the story data from the same articles. This script combines report entries based on URL, since it is the only constant identifier among the multiple title entries. Upon completion, the output of this script will contain a single field of all titles for a specific article, the unique story URL, and a sum of all story-level data across title entries.